[fix] connection fix
